third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/misc/resource-timing-sizes-tags-expected.txt

This is a testharness.js-based test.
FAIL PerformanceResourceTiming sizes tags test assert_equals: /svg/resources/greenSquare.svg?cb=8756076217654434 encodedBodySize expected 244 but got 0
Harness: the test ran to completion.
